Sukurtas atskiras .h failas šios programos antraštėm
