+++ global.h
@@ -30,14 +30,4 @@ using std::setprecision;
 using std::string_view;
 using std::sort;
 
-const int maxStud = 1000;
-struct Studentas {
-	string name = "Vardenis", surname = "Pavardenis";
-	vector <int> marks;
-	int markCount = 0;
-	int egzam = 0;
-	double average = 0;
-	double median = 0;
-};
-
 #endif
