+++ main.cpp
@@ -1,18 +1,5 @@
 #include "global.h"
-class Timer {
-private:
-	using hrClock = std::chrono::high_resolution_clock;
-	using durationDouble = std::chrono::duration<double>;
-	std::chrono::time_point<hrClock> start;
-public:
-	Timer() : start { hrClock::now() } {}
-	void reset() {
-	start = hrClock::now();
-	}
-	double elapsed() const {
-	return durationDouble (hrClock::now() - start).count();
-	}
-};
+#include "headers.h"
 
 int NumberCheck (int min, int max) {
 	int number;
